backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-core
langchain-groq
chromadb
pydantic
python-dotenv
pytesseract
pillow
